rename files for Writer classes
